uml entities' changes

+++ src/main/java/org/iesvdm/proyecto_plantquest/domain/Answer.java
@@ -35,5 +35,5 @@ public class Answer {
     private Question question;
 
     @Column(name = "propertyValue")
-    private double propertyValue;
+    private int propertyValue;
 }
